+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file.
 
 --------------------------------------
+
+####[1.0.1]
+
+[ADDED]	Simple LogWrapper above the android.util.Log class
+[ADDED] Fixes for signal protocol
+
 ####[1.0.0]
 
 Official release